Internally prices are doubles

+++ prr-core/src/prr/clients/BasePlan.java
@@ -14,16 +14,16 @@ public class BasePlan extends ClientPlan {
     @Serial
     private static final long serialVersionUID = 202222102220L;
 
-    private final int NORMAL_VOICE = 20;
-    private final int GOLD_VOICE = 10;
-    private final int PLATINUM_VOICE = 10;
+    private final double NORMAL_VOICE = 20;
+    private final double GOLD_VOICE = 10;
+    private final double PLATINUM_VOICE = 10;
 
-    private final int NORMAL_VIDEO = 30;
-    private final int GOLD_VIDEO = 20;
-    private final int PLATINUM_VIDEO = 10;
+    private final double NORMAL_VIDEO = 30;
+    private final double GOLD_VIDEO = 20;
+    private final double PLATINUM_VIDEO = 10;
 
     @Override
-    public long getPrice(TextCommunication text, Normal type, boolean isFriend) {
+    public double getPrice(TextCommunication text, Normal type, boolean isFriend) {
         if (text.getUnits() < 50)
             return 10;
         if (text.getUnits() < 100)
@@ -32,56 +32,56 @@ public class BasePlan extends ClientPlan {
     }
 
     @Override
-    public long getPrice(TextCommunication text, Gold type, boolean isFriend) {
+    public double getPrice(TextCommunication text, Gold type, boolean isFriend) {
         if (text.getUnits() < 100)
             return 10;
         return 2 * text.getUnits();
     }
 
     @Override
-    public long getPrice(TextCommunication text, Platinum type, boolean isFriend) {
+    public double getPrice(TextCommunication text, Platinum type, boolean isFriend) {
         if (text.getUnits() < 50)
             return 0;
         return 4;
     }
 
     @Override
-    public long getPrice(VoiceCommunication communication, Normal type, boolean isFriend) {
+    public double getPrice(VoiceCommunication communication, Normal type, boolean isFriend) {
         if (isFriend)
             return NORMAL_VOICE * 0.5;
         return NORMAL_VOICE;
     }
 
     @Override
-    public long getPrice(VoiceCommunication communication, Gold type, boolean isFriend) {
+    public double getPrice(VoiceCommunication communication, Gold type, boolean isFriend) {
         if (isFriend)
             return GOLD_VOICE * 0.5;
         return GOLD_VOICE;
     }
 
     @Override
-    public long getPrice(VoiceCommunication communication, Platinum type, boolean isFriend) {
+    public double getPrice(VoiceCommunication communication, Platinum type, boolean isFriend) {
         if (isFriend)
             return PLATINUM_VOICE * 0.5;
         return PLATINUM_VOICE;
     }
 
     @Override
-    public long getPrice(VideoCommunication communication, Normal type, boolean isFriend) {
+    public double getPrice(VideoCommunication communication, Normal type, boolean isFriend) {
         if (isFriend)
             return NORMAL_VIDEO * 0.5;
         return NORMAL_VIDEO;
     }
 
     @Override
-    public long getPrice(VideoCommunication communication, Gold type, boolean isFriend) {
+    public double getPrice(VideoCommunication communication, Gold type, boolean isFriend) {
         if (isFriend)
             return GOLD_VIDEO * 0.5;
         return GOLD_VIDEO;
     }
 
     @Override
-    public long getPrice(VideoCommunication communication, Platinum type, boolean isFriend) {
+    public double getPrice(VideoCommunication communication, Platinum type, boolean isFriend) {
         if (isFriend)
             return PLATINUM_VIDEO * 0.5;
         return PLATINUM_VIDEO;
